Consolidate basic header files from spacetime to modmesh
* Merge the contents of type.hpp and system.hpp in spacetime/ to modmesh/base.hpp
* Rename spacetime/type.hpp to spacetime/base_spacetime.hpp as a temporary alias
* Move spacetime/math.hpp to modmesh/math.hpp
