+++ cpp/modmesh/math.hpp
@@ -8,7 +8,7 @@
 #include <type_traits>
 #include <cmath>
 
-namespace spacetime
+namespace modmesh
 {
 
 namespace detail
@@ -34,6 +34,6 @@ inline constexpr T pow(T base)
     return detail::pow(base, std::integral_constant<size_t, N>());
 }
 
-} /* end namespace spacetime */
+} /* end namespace modmesh */
 
 // vim: set ff=unix fenc=utf8 et sw=4 ts=4 sts=4:
