+++ cpp/modmesh/spacetime/SolverBase_decl.hpp
@@ -8,8 +8,7 @@
 #include <memory>
 #include <vector>
 
-#include <modmesh/spacetime/system.hpp>
-#include <modmesh/spacetime/type.hpp>
+#include <modmesh/spacetime/base_spacetime.hpp>
 #include <modmesh/spacetime/Grid_decl.hpp>
 #include <modmesh/spacetime/Field_decl.hpp>
