+++ cpp/modmesh/spacetime/spacetime.hpp
@@ -8,9 +8,8 @@
 // FIXME: use modmesh/base.hpp
 #include <modmesh/modmesh.hpp>
 
-#include <modmesh/spacetime/system.hpp>
-#include <modmesh/spacetime/type.hpp>
-#include <modmesh/spacetime/math.hpp>
+#include <modmesh/spacetime/base_spacetime.hpp>
+#include "modmesh/math.hpp"
 #include <modmesh/spacetime/ElementBase.hpp>
 #include <modmesh/spacetime/Grid.hpp>
 #include <modmesh/spacetime/Celm.hpp>
