--- cpp/modmesh/spacetime/type.hpp
@@ -1,17 +1,0 @@
-#pragma once
-
-/*
- * Copyright (c) 2018, Yung-Yu Chen <yyc@solvcon.net>
- * BSD 3-Clause License, see COPYING
- */
-
-namespace spacetime
-{
-
-using real_type = double;
-using uint_type = uint32_t;
-using int_type = int32_t;
-
-} /* end namespace spacetime */
-
-// vim: set ff=unix fenc=utf8 et sw=4 ts=4 sts=4:
